requirements.txt
Flask==0.10.1
Flask-Login==0.2.9
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==1.0
Flask-Security==1.6.9
Flask-WTF==0.9.4
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
Pygments==1.6
SQLAlchemy==0.9.0
Sphinx==1.2
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.2
blinker==1.3
boto==2.24.0
docutils==0.11
flake8==2.1.0
gunicorn==18.0
ipython==1.1.0
itsdangerous==0.23
mccabe==0.2.1
passlib==1.6.2
pep8==1.4.6
psycopg2==2.5.1
py==1.4.20
pyflakes==0.7.3
pytest==2.5.2
python-dateutil==2.2
raven==4.0.3
requests==2.2.1
six==1.5.2
sphinx-rtd-theme==0.1.5
validictory==0.9.3
wsgiref==0.1.2
